HACK: add test harness and simulator/emulator
Feed in TGSI, and it compiles with both old and new compiler, and runs
resulting shader through emulator with same set of randomized inputs.  A
convenient way to test the new compiler.

Long term, maybe break the simulator out into a stand alone thing (which
takes ir3 asm as input??), since it is probably useful to keep even
after old compiler is removed.